% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECO637Data-package.R
\docType{package}
\name{ECO637Data-package}
\alias{ECO637Data}
\alias{ECO637Data-package}
\title{ECO637Data: Data package for ECO 637 (Empirical Health Economics at UNCG)}
\description{
Collects extracts from the 2012-2015 Medical Expenditure Panel Surveys for use in ECO 637.
}
\author{
\strong{Maintainer}: Martin Andersen \email{msander4@uncg.edu} (\href{https://orcid.org/0000-0001-5982-2541}{ORCID})

}
\keyword{internal}
