% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meps_2012.R, R/meps_2012_full.R
\docType{data}
\name{meps_2012}
\alias{meps_2012}
\alias{meps_2012_full}
\title{Extract of the 2012 Full Year Consolidated MEPS}
\format{
A tibble with 38974 rows and 107 (\code{meps_2012}) or 1883 variables (\code{meps_2012_full}).
\describe{
\item{duid}{DWELLING UNIT ID}
\item{pid}{PERSON NUMBER}
\item{dupersid}{PERSON ID (DUID + PID)}
\item{panel}{PANEL NUMBER}
\item{keyness}{PERSON KEY STATUS}
\item{insc1231}{INSCOPE STATUS ON 12/31/13}
\item{sex}{SEX}
\item{racever}{RACE QUESTION VERSION ASKED}
\item{racev1x}{RACE (EDITED/IMPUTED)}
\item{raceax}{ASIAN AMONG RACES RPTD (EDITED/IMPUTED)}
\item{racebx}{BLACK AMONG RACES RPTD (EDITED/IMPUTED)}
\item{racewx}{WHITE AMONG RACES RPTD (EDITED/IMPUTED)}
\item{racethx}{RACE/ETHNICITY (EDITED/IMPUTED)}
\item{hispanx}{HISPANIC ETHNICITY (EDITED/IMPUTED)}
\item{hispcat}{SPECIFIC HISPANIC ETHNICITY GROUP}
\item{edrecode}{EDUCATION RECODE (EDITED)}
\item{rthlth31}{PERCEIVED HEALTH STATUS - RD 3/1}
\item{rthlth42}{PERCEIVED HEALTH STATUS - RD 4/2}
\item{rthlth53}{PERCEIVED HEALTH STATUS - RD 5/3}
\item{mnhlth31}{PERCEIVED MENTAL HEALTH STATUS - RD 3/1}
\item{mnhlth42}{PERCEIVED MENTAL HEALTH STATUS - RD 4/2}
\item{mnhlth53}{PERCEIVED MENTAL HEALTH STATUS - RD 5/3}
\item{hibpdx}{HIGH BLOOD PRESSURE DIAG (>17)}
\item{bpmldx}{MULT DIAG HIGH BLOOD PRESS (>17)}
\item{chddx}{CORONARY HRT DISEASE DIAG (>17)}
\item{angidx}{ANGINA DIAGNOSIS (>17)}
\item{midx}{HEART ATTACK (MI) DIAG (>17)}
\item{ohrtdx}{OTHER HEART DISEASE DIAG (>17)}
\item{strkdx}{STROKE DIAGNOSIS (>17)}
\item{emphdx}{EMPHYSEMA DIAGNOSIS (>17)}
\item{chbron31}{CHRONC BRONCHITS LAST 12 MTHS (>17)-R3/1}
\item{chbron53}{CHRONC BRONCHITS LAST 12 MTHS (>17)-R5/3}
\item{choldx}{HIGH CHOLESTEROL DIAGNOSIS (>17)}
\item{cancerdx}{CANCER DIAGNOSIS (>17)}
\item{diabdx}{DIABETES DIAGNOSIS (>17)}
\item{jtpain31}{JOINT PAIN LAST 12 MONTHS (>17) - RD 3/1}
\item{jtpain53}{JOINT PAIN LAST 12 MONTHS (>17) - RD 5/3}
\item{arthdx}{ARTHRITIS DIAGNOSIS (>17)}
\item{arthtype}{TYPE OF ARTHRITIS DIAGNOSED (>17)}
\item{asthdx}{ASTHMA DIAGNOSIS}
\item{seatbe53}{WEARS SEAT BELT (>15) - RD 5/3}
\item{pcs42}{SAQ:PHY COMPONENT SUMMRY SF-12V2 IMPUTED}
\item{mcs42}{SAQ:MNT COMPONENT SUMMRY SF-12V2 IMPUTED}
}

A tibble with 38974 rows and 1883 variables
}
\source{
\url{https://meps.ahrq.gov/mepsweb/data_stats/download_data_files_detail.jsp?cboPufNumber=HC-155}
}
\usage{
meps_2012

meps_2012_full
}
\description{
Subset (or full file) of the 2012 Medical Expenditure Panel Survey full year consolidated data file (HC-155) for use in Empirical Health Economics (ECO 637). Full documentation can be found at the (MEPS website)[https://meps.ahrq.gov/data_stats/download_data_files_detail.jsp?cboPufNumber=HC-155].
}
\section{Functions}{
\itemize{
\item \code{meps_2012_full}: Extract of the 2012 Full Year Consolidated MEPS

Subset (or full file) of the 2012 Medical Expenditure Panel Survey full year consolidated data file (HC-155) for use in Empirical Health Economics (ECO 637). Full documentation can be found at the (MEPS website)[https://meps.ahrq.gov/data_stats/download_data_files_detail.jsp?cboPufNumber=HC-155].

}}
\seealso{
[meps_2012]
}
\keyword{datasets}
